[components] Change CLI from verb-noun to noun-verb
