+++ python_modules/libraries/dagster-dg/dagster_dg_tests/cli_tests/test_deployment_commands.py
@@ -1,0 +1,32 @@
+import os
+from pathlib import Path
+
+from dagster_dg.utils import ensure_dagster_dg_tests_import
+
+ensure_dagster_dg_tests_import()
+
+from dagster_dg_tests.utils import ProxyRunner, assert_runner_result
+
+# ########################
+# ##### GENERATE
+# ########################
+
+
+def test_generate_deployment_command_success() -> None:
+    with ProxyRunner.test() as runner, runner.isolated_filesystem():
+        result = runner.invoke("deployment", "generate", "foo")
+        assert_runner_result(result)
+        assert Path("foo").exists()
+        assert Path("foo/.github").exists()
+        assert Path("foo/.github/workflows").exists()
+        assert Path("foo/.github/workflows/dagster-cloud-deploy.yaml").exists()
+        assert Path("foo/dagster_cloud.yaml").exists()
+        assert Path("foo/code_locations").exists()
+
+
+def test_generate_deployment_command_already_exists_fails() -> None:
+    with ProxyRunner.test() as runner, runner.isolated_filesystem():
+        os.mkdir("foo")
+        result = runner.invoke("deployment", "generate", "foo")
+        assert_runner_result(result, exit_0=False)
+        assert "already exists" in result.output
