+++ python_modules/libraries/dagster-dg/dagster_dg_tests/test_cache.py
@@ -17,54 +17,54 @@ example_code_location = partial(isolated_example_code_location_bar, skip_venv=Tr
 
 def test_load_from_cache():
     with ProxyRunner.test(verbose=True) as runner, example_code_location(runner):
-        result = runner.invoke("list", "component-types")
+        result = runner.invoke("component-type", "list")
         assert_runner_result(result)
         assert "CACHE [miss]" in result.output
         assert "CACHE [write]" in result.output
-        result = runner.invoke("list", "component-types")
+        result = runner.invoke("component-type", "list")
         assert_runner_result(result)
         assert "CACHE [hit]" in result.output
 
 
 def test_cache_invalidation_uv_lock():
     with ProxyRunner.test(verbose=True) as runner, example_code_location(runner):
-        result = runner.invoke("list", "component-types")
+        result = runner.invoke("component-type", "list")
         assert_runner_result(result)
         assert "CACHE [miss]" in result.output
         assert "CACHE [write]" in result.output
 
         subprocess.run(["uv", "add", "dagster-components[dbt]", "dagster-dbt"], check=True)
 
-        result = runner.invoke("list", "component-types")
+        result = runner.invoke("component-type", "list")
         assert_runner_result(result)
         assert "CACHE [miss]" in result.output
 
 
 def test_cache_invalidation_modified_lib():
     with ProxyRunner.test(verbose=True) as runner, example_code_location(runner):
-        result = runner.invoke("list", "component-types")
+        result = runner.invoke("component-type", "list")
         assert_runner_result(result)
         assert "CACHE [miss]" in result.output
         assert "CACHE [write]" in result.output
 
-        result = runner.invoke("generate", "component-type", "my_component")
+        result = runner.invoke("component-type", "generate", "my_component")
         assert_runner_result(result)
 
-        result = runner.invoke("list", "component-types")
+        result = runner.invoke("component-type", "list")
         assert_runner_result(result)
         assert "CACHE [miss]" in result.output
 
 
 def test_cache_no_invalidation_modified_pkg():
     with ProxyRunner.test(verbose=True) as runner, example_code_location(runner):
-        result = runner.invoke("list", "component-types")
+        result = runner.invoke("component-type", "list")
         assert_runner_result(result)
         assert "CACHE [miss]" in result.output
         assert "CACHE [write]" in result.output
 
         Path("bar/submodule.py").write_text("print('hello')")
 
-        result = runner.invoke("list", "component-types")
+        result = runner.invoke("component-type", "list")
         assert_runner_result(result)
         assert "CACHE [hit]" in result.output
 
@@ -72,19 +72,19 @@ def test_cache_no_invalidation_modified_pkg():
 @pytest.mark.parametrize("with_command", [True, False])
 def test_cache_clear(with_command: bool):
     with ProxyRunner.test(verbose=True) as runner, example_code_location(runner):
-        result = runner.invoke("list", "component-types")
+        result = runner.invoke("component-type", "list")
         assert_runner_result(result)
         assert "CACHE [miss]" in result.output
         assert "CACHE [write]" in result.output
 
         if with_command:
-            result = runner.invoke("--clear-cache", "list", "component-types")
+            result = runner.invoke("--clear-cache", "component-type", "list")
             assert "CACHE [clear-all]" in result.output
         else:
             result = runner.invoke("--clear-cache")
             assert_runner_result(result)
             assert "CACHE [clear-all]" in result.output
-            result = runner.invoke("list", "component-types")
+            result = runner.invoke("component-type", "list")
 
         assert_runner_result(result)
         assert "CACHE [miss]" in result.output
@@ -100,7 +100,7 @@ def test_rebuild_component_registry_success():
         assert_runner_result(result)
         assert "CACHE [clear-key]" in result.output
 
-        result = runner.invoke("list", "component-types")
+        result = runner.invoke("component-type", "list")
         assert_runner_result(result)
         assert "CACHE [hit]" in result.output
 
@@ -114,7 +114,7 @@ def test_rebuild_component_registry_fails_outside_code_location():
 
 def test_rebuild_component_registry_fails_with_subcommand():
     with ProxyRunner.test(verbose=True) as runner, example_code_location(runner):
-        result = runner.invoke("--rebuild-component-registry", "list", "component-types")
+        result = runner.invoke("--rebuild-component-registry", "component-type", "list")
         assert_runner_result(result, exit_0=False)
         assert "Cannot specify --rebuild-component-registry with a subcommand." in result.output
 
@@ -138,6 +138,6 @@ def test_cache_disabled():
         ProxyRunner.test(verbose=True, disable_cache=True) as runner,
         example_code_location(runner),
     ):
-        result = runner.invoke("list", "component-types")
+        result = runner.invoke("component-type", "list")
         assert_runner_result(result)
         assert "CACHE" not in result.output
